Add and use string macros to toggle isolates
In versions of Node before 11 String->Write* and String->Concat did not
require the isolate as an argument. The new macros STRING_WRITE,
STRING_WRITEUTF8 and STRING_CONCAT handle calling the correct method
and excluding the isolate if the Node version does not require it.

This allows compatibility for Node versions 3 through 16